refactor: mock utils
