+++ src/renderer/__mocks__/account-mocks.ts
@@ -8,6 +8,15 @@ import type {
 } from '../types';
 import { mockGitifyUser } from './user-mocks';
 
+export const mockGitHubAppAccount: Account = {
+  platform: 'GitHub Cloud',
+  method: 'GitHub App',
+  token: 'token-987654321' as Token,
+  hostname: Constants.DEFAULT_AUTH_OPTIONS.hostname,
+  user: mockGitifyUser,
+  hasRequiredScopes: true,
+};
+
 export const mockPersonalAccessTokenAccount: Account = {
   platform: 'GitHub Cloud',
   method: 'Personal Access Token',
@@ -20,7 +29,7 @@ export const mockPersonalAccessTokenAccount: Account = {
 export const mockOAuthAccount: Account = {
   platform: 'GitHub Enterprise Server',
   method: 'OAuth App',
-  token: '1234568790' as Token,
+  token: 'token-1234568790' as Token,
   hostname: 'github.gitify.io' as Hostname,
   user: mockGitifyUser,
   hasRequiredScopes: true,
@@ -39,24 +48,13 @@ export const mockGitHubCloudAccount: Account = {
 export const mockGitHubEnterpriseServerAccount: Account = {
   platform: 'GitHub Enterprise Server',
   method: 'Personal Access Token',
-  token: '1234568790' as Token,
+  token: 'token-1234568790' as Token,
   hostname: 'github.gitify.io' as Hostname,
   user: mockGitifyUser,
   hasRequiredScopes: true,
 };
 
-export const mockGitHubAppAccount: Account = {
-  platform: 'GitHub Cloud',
-  method: 'GitHub App',
-  token: '987654321' as Token,
-  hostname: Constants.DEFAULT_AUTH_OPTIONS.hostname,
-  user: mockGitifyUser,
-  hasRequiredScopes: true,
-};
-
-export function createMockAccountWithError(
-  error: GitifyError,
-): AccountNotifications {
+export function mockAccountWithError(error: GitifyError): AccountNotifications {
   return {
     account: mockGitHubCloudAccount,
     notifications: [],
