+++ src/renderer/__mocks__/user-mocks.ts
@@ -1,6 +1,4 @@
 import type { GitifyNotificationUser, GitifyUser, Link } from '../types';
-import type { AuthorFieldsFragment } from '../utils/api/graphql/generated/graphql';
-import type { RawUser } from '../utils/api/types';
 
 export const mockGitifyUser: GitifyUser = {
   login: 'octocat',
@@ -9,18 +7,7 @@ export const mockGitifyUser: GitifyUser = {
   avatar: 'https://avatars.githubusercontent.com/u/583231?v=4' as Link,
 };
 
-export function createPartialMockUser(login: string): RawUser {
-  const mockUser: Partial<RawUser> = {
-    login: login,
-    html_url: `https://github.com/${login}` as Link,
-    avatar_url: 'https://avatars.githubusercontent.com/u/583231?v=4' as Link,
-    type: 'User',
-  };
-
-  return mockUser as RawUser;
-}
-
-export function createMockGitifyNotificationUser(
+export function mockGitifyNotificationUser(
   login: string,
 ): GitifyNotificationUser {
   return {
@@ -29,14 +16,4 @@ export function createMockGitifyNotificationUser(
     avatarUrl: 'https://avatars.githubusercontent.com/u/583231?v=4' as Link,
     type: 'User',
   };
-}
-
-/**
- * Creates a mock author for use in GraphQL response mocks.
- */
-export function createMockGraphQLAuthor(login: string): AuthorFieldsFragment {
-  return {
-    ...createMockGitifyNotificationUser(login),
-    __typename: 'User',
-  } as AuthorFieldsFragment;
 }
