+++ src/renderer/components/filters/StateFilter.test.tsx
@@ -1,11 +1,11 @@
 import { renderWithAppContext } from '../../__helpers__/test-utils';
-import { mockAccountNotifications } from '../../__mocks__/notifications-mocks';
+import { mockMultipleAccountNotifications } from '../../__mocks__/notifications-mocks';
 import { StateFilter } from './StateFilter';
 
 describe('renderer/components/filters/StateFilter.tsx', () => {
   it('should render itself & its children', () => {
     const tree = renderWithAppContext(<StateFilter />, {
-      notifications: mockAccountNotifications,
+      notifications: mockMultipleAccountNotifications,
     });
 
     expect(tree).toMatchSnapshot();
