+++ src/renderer/components/filters/SubjectTypeFilter.test.tsx
@@ -1,11 +1,11 @@
 import { renderWithAppContext } from '../../__helpers__/test-utils';
-import { mockAccountNotifications } from '../../__mocks__/notifications-mocks';
+import { mockMultipleAccountNotifications } from '../../__mocks__/notifications-mocks';
 import { SubjectTypeFilter } from './SubjectTypeFilter';
 
 describe('renderer/components/filters/SubjectTypeFilter.tsx', () => {
   it('should render itself & its children', () => {
     const tree = renderWithAppContext(<SubjectTypeFilter />, {
-      notifications: mockAccountNotifications,
+      notifications: mockMultipleAccountNotifications,
     });
 
     expect(tree).toMatchSnapshot();
