+++ src/renderer/components/metrics/MetricGroup.test.tsx
@@ -1,13 +1,13 @@
 import { renderWithAppContext } from '../../__helpers__/test-utils';
+import { mockGitifyNotification } from '../../__mocks__/notifications-mocks';
 import { mockSettings } from '../../__mocks__/state-mocks';
 import type { GitifyMilestone } from '../../types';
-import { mockSingleNotification } from '../../utils/api/__mocks__/response-mocks';
 import { MetricGroup } from './MetricGroup';
 
 describe('renderer/components/metrics/MetricGroup.tsx', () => {
   describe('showPills disabled', () => {
     it('should not render any pills when showPills is disabled', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
 
       const props = {
         notification: mockNotification,
@@ -22,7 +22,7 @@ describe('renderer/components/metrics/MetricGroup.tsx', () => {
 
   describe('linked issue pills', () => {
     it('should render issues pill when linked to one issue/pr', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
       mockNotification.subject.linkedIssues = ['#1'];
 
       const props = {
@@ -34,7 +34,7 @@ describe('renderer/components/metrics/MetricGroup.tsx', () => {
     });
 
     it('should render issues pill when linked to multiple issues/prs', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
       mockNotification.subject.linkedIssues = ['#1', '#2'];
 
       const props = {
@@ -48,7 +48,7 @@ describe('renderer/components/metrics/MetricGroup.tsx', () => {
 
   describe('comment pills', () => {
     it('should render when no comments', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
       mockNotification.subject.comments = null;
 
       const props = {
@@ -60,7 +60,7 @@ describe('renderer/components/metrics/MetricGroup.tsx', () => {
     });
 
     it('should render when 1 comment', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
       mockNotification.subject.comments = 1;
 
       const props = {
@@ -72,7 +72,7 @@ describe('renderer/components/metrics/MetricGroup.tsx', () => {
     });
 
     it('should render when more than 1 comments', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
       mockNotification.subject.comments = 2;
 
       const props = {
@@ -86,7 +86,7 @@ describe('renderer/components/metrics/MetricGroup.tsx', () => {
 
   describe('label pills', () => {
     it('should render labels pill', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
       mockNotification.subject.labels = ['enhancement', 'good-first-issue'];
 
       const props = {
@@ -100,7 +100,7 @@ describe('renderer/components/metrics/MetricGroup.tsx', () => {
 
   describe('milestone pills', () => {
     it('should render open milestone pill', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
       mockNotification.subject.milestone = {
         title: 'Milestone 1',
         state: 'OPEN',
@@ -115,7 +115,7 @@ describe('renderer/components/metrics/MetricGroup.tsx', () => {
     });
 
     it('should render closed milestone pill', async () => {
-      const mockNotification = mockSingleNotification;
+      const mockNotification = mockGitifyNotification;
       mockNotification.subject.milestone = {
         title: 'Milestone 1',
         state: 'CLOSED',
