+++ src/renderer/hooks/useNotifications.test.ts
@@ -3,8 +3,8 @@ import { act, renderHook, waitFor } from '@testing-library/react';
 import axios, { AxiosError } from 'axios';
 import nock from 'nock';
 
+import { mockGitifyNotification } from '../__mocks__/notifications-mocks';
 import { mockAuth, mockSettings, mockState } from '../__mocks__/state-mocks';
-import { mockSingleNotification } from '../utils/api/__mocks__/response-mocks';
 import { Errors } from '../utils/errors';
 import * as logger from '../utils/logger';
 import { useNotifications } from './useNotifications';
@@ -21,10 +21,10 @@ describe('renderer/hooks/useNotifications.ts', () => {
     rendererLogErrorSpy.mockReset();
 
     // Reset mock notification state between tests since it's mutated
-    mockSingleNotification.unread = true;
+    mockGitifyNotification.unread = true;
   });
 
-  const id = mockSingleNotification.id;
+  const id = mockGitifyNotification.id;
 
   describe('fetchNotifications', () => {
     it('should fetch non-detailed notifications with success', async () => {
@@ -213,7 +213,7 @@ describe('renderer/hooks/useNotifications.ts', () => {
 
       act(() => {
         result.current.markNotificationsAsRead(mockState, [
-          mockSingleNotification,
+          mockGitifyNotification,
         ]);
       });
 
@@ -233,7 +233,7 @@ describe('renderer/hooks/useNotifications.ts', () => {
 
       act(() => {
         result.current.markNotificationsAsRead(mockState, [
-          mockSingleNotification,
+          mockGitifyNotification,
         ]);
       });
 
@@ -256,7 +256,7 @@ describe('renderer/hooks/useNotifications.ts', () => {
 
       act(() => {
         result.current.markNotificationsAsDone(mockState, [
-          mockSingleNotification,
+          mockGitifyNotification,
         ]);
       });
 
@@ -276,7 +276,7 @@ describe('renderer/hooks/useNotifications.ts', () => {
 
       act(() => {
         result.current.markNotificationsAsDone(mockState, [
-          mockSingleNotification,
+          mockGitifyNotification,
         ]);
       });
 
@@ -308,7 +308,7 @@ describe('renderer/hooks/useNotifications.ts', () => {
       act(() => {
         result.current.unsubscribeNotification(
           mockState,
-          mockSingleNotification,
+          mockGitifyNotification,
         );
       });
 
@@ -341,7 +341,7 @@ describe('renderer/hooks/useNotifications.ts', () => {
               markAsDoneOnUnsubscribe: true,
             },
           },
-          mockSingleNotification,
+          mockGitifyNotification,
         );
       });
 
@@ -368,7 +368,7 @@ describe('renderer/hooks/useNotifications.ts', () => {
       act(() => {
         result.current.unsubscribeNotification(
           mockState,
-          mockSingleNotification,
+          mockGitifyNotification,
         );
       });
