+++ src/renderer/utils/api/client.test.ts
@@ -1,7 +1,10 @@
 import axios, { type AxiosPromise, type AxiosResponse } from 'axios';
 
 import { mockGitHubCloudAccount } from '../../__mocks__/account-mocks';
-import { createPartialMockNotification } from '../../__mocks__/notifications-mocks';
+import {
+  mockGitHubCloudGitifyNotifications,
+  mockPartialGitifyNotification,
+} from '../../__mocks__/notifications-mocks';
 import { mockToken } from '../../__mocks__/state-mocks';
 import { Constants } from '../../constants';
 import type { Hostname, Link, SettingsState, Token } from '../../types';
@@ -10,7 +13,6 @@ import {
   mockAuthHeaders,
   mockNonCachedAuthHeaders,
 } from './__mocks__/request-mocks';
-import { mockGitHubNotifications } from './__mocks__/response-mocks';
 import {
   fetchAuthenticatedUserDetails,
   fetchDiscussionByNumber,
@@ -209,7 +211,7 @@ describe('renderer/utils/api/client.ts', () => {
       'performGraphQLRequest',
     );
 
-    const mockNotification = createPartialMockNotification({
+    const mockNotification = mockPartialGitifyNotification({
       title: 'Some discussion',
       url: 'https://api.github.com/repos/gitify-app/gitify/discussion/123' as Link,
       type: 'Discussion',
@@ -224,7 +226,7 @@ describe('renderer/utils/api/client.ts', () => {
 
     expect(performGraphQLRequestSpy).toHaveBeenCalledWith(
       'https://api.github.com/graphql',
-      mockToken,
+      mockNotification.account.token,
       FetchDiscussionByNumberDocument,
       {
         owner: mockNotification.repository.owner.login,
@@ -244,7 +246,7 @@ describe('renderer/utils/api/client.ts', () => {
       'performGraphQLRequest',
     );
 
-    const mockNotification = createPartialMockNotification({
+    const mockNotification = mockPartialGitifyNotification({
       title: 'Some issue',
       url: 'https://api.github.com/repos/gitify-app/gitify/issues/123' as Link,
       type: 'Issue',
@@ -259,7 +261,7 @@ describe('renderer/utils/api/client.ts', () => {
 
     expect(performGraphQLRequestSpy).toHaveBeenCalledWith(
       'https://api.github.com/graphql',
-      mockToken,
+      mockNotification.account.token,
       FetchIssueByNumberDocument,
       {
         owner: mockNotification.repository.owner.login,
@@ -277,7 +279,7 @@ describe('renderer/utils/api/client.ts', () => {
       'performGraphQLRequest',
     );
 
-    const mockNotification = createPartialMockNotification({
+    const mockNotification = mockPartialGitifyNotification({
       title: 'Some pull request',
       url: 'https://api.github.com/repos/gitify-app/gitify/pulls/123' as Link,
       type: 'PullRequest',
@@ -292,7 +294,7 @@ describe('renderer/utils/api/client.ts', () => {
 
     expect(performGraphQLRequestSpy).toHaveBeenCalledWith(
       'https://api.github.com/graphql',
-      mockToken,
+      mockNotification.account.token,
       FetchPullRequestByNumberDocument,
       {
         owner: mockNotification.repository.owner.login,
@@ -313,7 +315,7 @@ describe('renderer/utils/api/client.ts', () => {
         'performGraphQLRequestString',
       );
 
-      const mockNotification = createPartialMockNotification({
+      const mockNotification = mockPartialGitifyNotification({
         title: 'Some commit',
         url: 'https://api.github.com/repos/gitify-app/gitify/commit/123' as Link,
         type: 'Commit',
@@ -356,7 +358,7 @@ describe('renderer/utils/api/client.ts', () => {
         headers: {},
       } as ExecutionResultWithHeaders<unknown>);
 
-      await fetchNotificationDetailsForList(mockGitHubNotifications);
+      await fetchNotificationDetailsForList(mockGitHubCloudGitifyNotifications);
 
       expect(performGraphQLRequestStringSpy).toHaveBeenCalledWith(
         'https://api.github.com/graphql',
