+++ src/renderer/utils/notifications/filters/filter.test.ts
@@ -1,4 +1,4 @@
-import { createPartialMockNotification } from '../../../__mocks__/notifications-mocks';
+import { mockPartialGitifyNotification } from '../../../__mocks__/notifications-mocks';
 import { mockSettings } from '../../../__mocks__/state-mocks';
 import { defaultSettings } from '../../../context/defaults';
 import type {
@@ -20,7 +20,7 @@ describe('renderer/utils/notifications/filters/filter.ts', () => {
 
   describe('filterNotifications', () => {
     const mockNotifications = [
-      createPartialMockNotification(
+      mockPartialGitifyNotification(
         {
           title: 'User authored notification',
           user: {
@@ -34,13 +34,14 @@ describe('renderer/utils/notifications/filters/filter.ts', () => {
         {
           owner: {
             login: 'gitify-app',
-            avatarUrl: 'https://avatars.githubusercontent.com/u/1' as Link,
+            avatarUrl:
+              'https://avatars.githubusercontent.com/u/133795385?s=200&v=4' as Link,
             type: 'Organization',
           } as GitifyOwner,
           fullName: 'gitify-app/gitify',
         },
       ),
-      createPartialMockNotification(
+      mockPartialGitifyNotification(
         {
           title: 'Bot authored notification',
           user: {
@@ -54,7 +55,8 @@ describe('renderer/utils/notifications/filters/filter.ts', () => {
         {
           owner: {
             login: 'github',
-            avatarUrl: 'https://avatars.githubusercontent.com/u/2' as Link,
+            avatarUrl:
+              'https://avatars.githubusercontent.com/u/133795385?s=200&v=4' as Link,
             type: 'Organization',
           } as GitifyOwner,
           fullName: 'github/github',
