+++ src/renderer/utils/notifications/handlers/repositoryVulnerabilityAlert.test.ts
@@ -1,10 +1,10 @@
-import { createPartialMockNotification } from '../../../__mocks__/notifications-mocks';
+import { mockPartialGitifyNotification } from '../../../__mocks__/notifications-mocks';
 import type { GitifyNotification, Link } from '../../../types';
 import { repositoryVulnerabilityAlertHandler } from './repositoryVulnerabilityAlert';
 
 describe('renderer/utils/notifications/handlers/repositoryVulnerabilityAlert.ts', () => {
   it('iconType', () => {
-    const mockNotification = createPartialMockNotification({
+    const mockNotification = mockPartialGitifyNotification({
       type: 'RepositoryVulnerabilityAlert',
     });
