+++ src/renderer/utils/notifications/remove.test.ts
@@ -1,7 +1,9 @@
 import { mockGitHubEnterpriseServerAccount } from '../../__mocks__/account-mocks';
-import { mockSingleAccountNotifications } from '../../__mocks__/notifications-mocks';
+import {
+  mockGitifyNotification,
+  mockSingleAccountNotifications,
+} from '../../__mocks__/notifications-mocks';
 import { mockSettings } from '../../__mocks__/state-mocks';
-import { mockSingleNotification } from '../api/__mocks__/response-mocks';
 import { removeNotificationsForAccount } from './remove';
 
 describe('renderer/utils/remove.ts', () => {
@@ -11,7 +13,7 @@ describe('renderer/utils/remove.ts', () => {
     const result = removeNotificationsForAccount(
       mockSingleAccountNotifications[0].account,
       { ...mockSettings, delayNotificationState: false },
-      [mockSingleNotification],
+      [mockGitifyNotification],
       mockSingleAccountNotifications,
     );
 
@@ -24,7 +26,7 @@ describe('renderer/utils/remove.ts', () => {
     const result = removeNotificationsForAccount(
       mockSingleAccountNotifications[0].account,
       { ...mockSettings, delayNotificationState: true },
-      [mockSingleNotification],
+      [mockGitifyNotification],
       mockSingleAccountNotifications,
     );
 
@@ -40,7 +42,7 @@ describe('renderer/utils/remove.ts', () => {
       { ...mockSettings, delayNotificationState: true },
       [
         {
-          ...mockSingleNotification,
+          ...mockGitifyNotification,
           id: 'non-existent-id',
         },
       ],
@@ -71,7 +73,7 @@ describe('renderer/utils/remove.ts', () => {
     const result = removeNotificationsForAccount(
       mockGitHubEnterpriseServerAccount, // Different account
       { ...mockSettings, delayNotificationState: false },
-      [mockSingleNotification],
+      [mockGitifyNotification],
       mockSingleAccountNotifications,
     );
